주별로 다시 정리

+++ python/manage_climate_data.py
@@ -6,6 +6,8 @@ path = "C:/Users/csi2/Desktop/bee-colony-RnE/data/climate"
 
 def unzip_file():
     for state in os.listdir(path):
+        if state.endswith('.py') or state.endswith('.md'):
+            continue
         for year in os.listdir(path+'/'+state):
             if year.endswith('.zip'):
                 with zipfile.ZipFile(path+'/'+state+'/'+year, 'r') as zip_ref:
@@ -17,6 +19,8 @@ def unzip_file():
 
 def classify_file():
     for state in os.listdir(path):
+        if state.endswith('.py') or state.endswith('.md'):
+            continue
         if not os.path.exists(path+'/'+state+'/all reports'):
             os.makedirs(path+'/'+state+'/all reports')
         if not os.path.exists(path+'/'+state+'/daily summury'):
@@ -30,6 +34,12 @@ def classify_file():
 
 def clean_directory():
     for state in os.listdir(path):
+        if state.endswith('.py') or state.endswith('.md'):
+            continue
         for folder in os.listdir(path+'/'+state):
             if folder != 'all reports' and folder != 'daily summury':
                 shutil.rmtree(path+'/'+state+'/'+folder)
+
+unzip_file()
+classify_file()
+clean_directory()
